+++ package.json
@@ -40,7 +40,7 @@
     "@emotion/react": "11.7.1",
     "@emotion/styled": "11.6.0",
     "@mui/lab": "5.0.0-alpha.60",
-    "@mui/material": "5.4.1",
+    "@mui/material": "5.4.2",
     "@react-spring/web": "9.4.3",
     "@storybook/addon-docs": "6.4.19",
     "@testing-library/jest-dom": "5.16.2",
@@ -50,10 +50,10 @@
     "@vercel/fetch": "6.1.1",
     "autoprefixer": "10.4.2",
     "autosuggest-highlight": "3.2.0",
-    "axios": "0.25.0",
+    "axios": "0.26.0",
     "body-scroll-lock": "3.1.5",
     "classnames": "2.3.1",
-    "cookie": "0.4.1",
+    "cookie": "0.4.2",
     "cross-env": "7.0.3",
     "email-validator": "2.0.4",
     "immutability-helper": "3.1.1",
@@ -67,7 +67,7 @@
     "next-seo": "4.28.1",
     "next-share": "0.13.0",
     "next-sitemap": "1.9.9",
-    "next-themes": "0.0.14",
+    "next-themes": "0.0.15",
     "postcss": "8.4.6",
     "postcss-nesting": "8.0.1",
     "react": "17.0.2",
@@ -99,7 +99,7 @@
     "@storybook/manager-webpack5": "6.4.19",
     "@storybook/react": "6.4.19",
     "@types/body-scroll-lock": "2.6.1",
-    "@types/cookie": "0.4.0",
+    "@types/cookie": "0.4.1",
     "@types/js-cookie": "2.2.6",
     "@types/lodash.debounce": "4.0.6",
     "@types/lodash.random": "3.2.6",
